pyproject.toml
[project]
name = "gossip"
version = "0.1.0"
description = "Open-source social utility bot that keeps group chats alive"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # LLM
    "anthropic>=0.39.0",

    # Google APIs
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.160.0",

    # Database
    "aiosqlite>=0.20.0",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "croniter>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["gossip*", "gossip_tools*", "portal*"]

[tool.ruff]
line-length = 100
target-version = "py311"
